feat:抽离表单中的变量，新增VariablePanel;优化添加的展开项逻辑
