+++ app/renderer/src/main/src/pages/fuzzer/HttpQueryAdvancedConfig/FuzzerConfigPanels.tsx
@@ -1,21 +1,29 @@
-import React from "react"
+import React, {useRef} from "react"
 import styles from "./HttpQueryAdvancedConfig.module.scss"
 import YakitCollapse from "@/components/yakitUI/YakitCollapse/YakitCollapse"
 import {YakitButton} from "@/components/yakitUI/YakitButton/YakitButton"
 import {useCreation, useMemoizedFn} from "ahooks"
-import {Divider, Form} from "antd"
+import {Descriptions, Divider, Form} from "antd"
 import {HollowLightningBoltIcon} from "@/assets/newIcon"
 import {MatchingAndExtraction} from "../MatcherAndExtractionCard/MatcherAndExtractionCardType"
 import {YakitRadioButtons} from "@/components/yakitUI/YakitRadioButtons/YakitRadioButtons"
 import {
     ColorSelect,
+    defMatcherAndExtractionCode,
     filterModeOptions,
     matchersConditionOptions
 } from "../MatcherAndExtractionCard/MatcherAndExtractionCard"
 import {ExtractorsList, MatchersList} from "./HttpQueryAdvancedConfig"
 import {AdvancedConfigValueProps} from "./HttpQueryAdvancedConfigType"
+import {StringToUint8Array} from "@/utils/str"
+import {showYakitModal} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
+import {CopyableField} from "@/utils/inputUtil"
+import {yakitFailed, yakitNotify} from "@/utils/notification"
+import {VariableList} from "@/pages/httpRequestBuilder/HTTPRequestBuilder"
+import {OutlinePlusIcon} from "@/assets/icon/outline"
 
 const {YakitPanel} = YakitCollapse
+const {ipcRenderer} = window.require("electron")
 
 interface MatchersPanelProps {
     onAddMatchingAndExtractionCard: (type: MatchingAndExtraction) => void
@@ -221,6 +229,151 @@ export const ExtractorsPanel: React.FC<ExtractorsPanelProps> = React.memo((props
                     onRemove={onRemoveExtractors}
                     onEdit={onEditExtractors}
                 />
+            </YakitPanel>
+        </>
+    )
+})
+const variableModeOptions = [
+    {
+        value: "nuclei-dsl",
+        label: "nuclei"
+    },
+    {
+        value: "fuzztag",
+        label: "fuzztag"
+    },
+    {
+        value: "raw",
+        label: "raw"
+    }
+]
+interface VariablePanelProps {
+    defaultHttpResponse: string
+    onAdd: (s: string) => void
+}
+export const VariablePanel: React.FC<VariablePanelProps> = React.memo((props) => {
+    const {defaultHttpResponse, onAdd, ...restProps} = props
+    const form = Form.useFormInstance()
+    const params = Form.useWatch("params", form) || []
+
+    const variableRef = useRef<any>()
+
+    const onReset = useMemoizedFn(() => {
+        form.setFieldsValue({
+            params: [{Key: "", Value: "", Type: "raw"}]
+        })
+    })
+    /** @description 变量预览 */
+    const onRenderVariables = useMemoizedFn((e: React.MouseEvent<HTMLElement, MouseEvent>) => {
+        e.stopPropagation()
+        ipcRenderer
+            .invoke("RenderVariables", {
+                Params: form.getFieldValue("params") || [],
+                HTTPResponse: StringToUint8Array(defaultHttpResponse || defMatcherAndExtractionCode)
+            })
+            .then((rsp: {Results: {Key: string; Value: string}[]}) => {
+                showYakitModal({
+                    title: "渲染后变量内容",
+                    footer: <></>,
+                    content: (
+                        <div className={styles["render-variables-modal-content"]}>
+                            <Descriptions size={"small"} column={1} bordered={true}>
+                                {rsp.Results.filter((i) => {
+                                    return !(i.Key === "" && i.Value === "")
+                                }).map((data, index) => {
+                                    return (
+                                        <Descriptions.Item
+                                            label={<CopyableField text={data.Key} maxWidth={100} />}
+                                            key={`${data.Key}-${data.Value}`}
+                                        >
+                                            <CopyableField text={data.Value} maxWidth={300} />
+                                        </Descriptions.Item>
+                                    )
+                                })}
+                            </Descriptions>
+                        </div>
+                    )
+                })
+            })
+            .catch((err) => {
+                yakitNotify("error", "预览失败:" + err)
+            })
+    })
+    const onAddPrams = useMemoizedFn(() => {
+        onAdd("设置变量")
+        const index = params.findIndex((ele: {Key: string; Value: string}) => !ele || (!ele.Key && !ele.Value))
+        if (index === -1) {
+            form.setFieldsValue({
+                params: [...params, {Key: "", Value: "", Type: "raw"}]
+            })
+            variableRef.current?.setVariableActiveKey([
+                ...(variableRef.current?.variableActiveKey || []),
+                `${params?.length || 0}`
+            ])
+        } else {
+            yakitFailed(`请将已添加【变量${index}】设置完成后再进行添加`)
+        }
+    })
+    const onRemove = useMemoizedFn((index: number) => {
+        const newParams = params.filter((_, i) => i !== index)
+        form.setFieldsValue({
+            params: [...newParams]
+        })
+    })
+    return (
+        <>
+            <YakitPanel
+                {...restProps}
+                header='设置变量'
+                key='设置变量'
+                extra={
+                    <>
+                        <YakitButton
+                            type='text'
+                            colors='danger'
+                            onClick={(e) => {
+                                e.stopPropagation()
+                                onReset()
+                            }}
+                            size='small'
+                        >
+                            重置
+                        </YakitButton>
+                        <Divider type='vertical' style={{margin: 0}} />
+                        <YakitButton type='text' onClick={onRenderVariables} size='small'>
+                            预览
+                        </YakitButton>
+                        <Divider type='vertical' style={{margin: 0}} />
+                        <YakitButton
+                            type='text'
+                            onClick={(e) => {
+                                e.stopPropagation()
+                                onAddPrams()
+                            }}
+                            className={styles["btn-padding-right-0"]}
+                            size='small'
+                        >
+                            添加
+                            <OutlinePlusIcon />
+                        </YakitButton>
+                    </>
+                }
+            >
+                <VariableList
+                    ref={variableRef}
+                    field='params'
+                    onDel={onRemove}
+                    extra={(i, info) => (
+                        <Form.Item name={[info.name, "Type"]} noStyle wrapperCol={{span: 24}}>
+                            <YakitRadioButtons
+                                style={{marginLeft: 4}}
+                                buttonStyle='solid'
+                                options={variableModeOptions}
+                                size={"small"}
+                            />
+                        </Form.Item>
+                    )}
+                ></VariableList>
             </YakitPanel>
         </>
     )
